minor fix in the constructor of Vector class

+++ src/Game/testing/Vector.java
@@ -3,7 +3,7 @@ package Game.testing;
 public class Vector {
     public double x, y;
 
-    public Vector(int x, int y) {
+    public Vector(double x, double y) {
         this.x = x;
         this.y = y;
     }
